file-tree explorer, bar and tabs, tmux integration and remaps
Added barbar.nvim for a visual hint of opened buffers/tabs in neovim;
Added neovim-neo-tree for a visual representation of the working
directory, useful for things like directory based routing;
Added better integration with tmux;
Created file for custom remaps.
Created TODO list for things that need to be added to the setup.